+++ src/ModularPipelines.Build/Settings/CodeCovSettings.cs
@@ -1,0 +1,6 @@
+﻿namespace ModularPipelines.Build.Settings;
+
+public class CodeCovSettings
+{
+    public string? Token { get; init; }
+}