+++ src/ModularPipelines.Build/ReleaseNotes.md
@@ -1,1 +1,2 @@
 *   StyleCop Analyzers to enforce code quality
+*   Upload Code Coverage results to CodeCov